LPS-203968 Move portal-impl PortletPreferencesModelListener logic to portal-configuration-module-configuration PortletPreferencesModelListener
